+++ src/DependencyInjection/Configuration.php
@@ -27,7 +27,11 @@ use Symfony\Component\Config\Definition\ConfigurationInterface;
 class Configuration implements ConfigurationInterface
 {
     /**
-     * {@inheritdoc}
+     * @psalm-suppress PossiblyNullReference, PossiblyUndefinedMethod
+     *
+     * @see https://github.com/psalm/psalm-plugin-symfony/issues/174
+     *
+     * @return TreeBuilder
      */
     public function getConfigTreeBuilder()
     {