+++ src/Twig/Extension/LocaleExtension.php
@@ -35,7 +35,7 @@ class LocaleExtension extends AbstractExtension
     }
 
     /**
-     * {@inheritdoc}
+     * @return TwigFilter[]
      */
     public function getFilters()
     {